server/requirements.txt
flask==1.1.2
gunicorn==19.7.1
pyOpenSSL==19.0.0
